NatureCounter/DjangoBackend/pyproject.toml
[tool.poetry]
name = "nature counter by crowddoing"
version = "0.0.1"
description = ""
authors = ["Meena Bharani <ns.BharaniMeena@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pymongo = "3.12.1"
djongo = "^1.3.6"
pyrebase = "^3.0.27"
psycopg2 = "^2.9.5"
scrapy = "^2.7.1"
autopep8 = "^2.0.1"
pep8 = "^1.7.1"
pylint = "^2.15.10"
pylint-django = "^2.5.3"
pillow = "^9.4.0"
sqlparse = "0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
